reduced memory size of JSON nodes
I’ve combined the tag and the string node string length into a single
uint16_t field by defining getters and setters.  This relies on the
compiler representing enums with an integer starting from zero. I
currently have 14 different tags so I have used a 5 bit field for the
tag, leaving 11 bits for strings. This limits  strings to 2048 bytes
each. This should be plenty on the Arduino Uno which has 2KB RAM, but
could be revisited on the ESP8266 and other devices with more RAM.