Making it easier to compile on Arduino IDE
Using PRINT as a macro for cout to facilitate sed script that
transforms this to Serial.print.  The arduino.sh script copies files
from the project directory (change to match your setup) and should be
run from your Arduino library folder for this server. I still have
problems with Serial.print not knowing how to deal with const unsigned
char *